Moved Builder States into dedicated folder
