+++ src/lib/dragon/index.ts
@@ -1,9 +1,3 @@
-export const BUILDER_STATES = ['LOADING', 'WELCOME', 'DISPLAY', 'EDIT', 'DEBUG'] as const;
-export type BuilderState = (typeof BUILDER_STATES)[number];
-export function stringToBuilderState(state_string: string): BuilderState | undefined {
-	return BUILDER_STATES.find((state) => state === state_string);
-}
-
 export const AGES = ['wyrmling', 'young', 'adult', 'ancient'] as const;
 export type Age = (typeof AGES)[number];
